Add splashes for multiple commands; build from SVG
For the UBports Recovery, we prefer to be able to differentiate between
an update boot and a no-command boot. So, add an 'update' and 'waiting'
splash to denote each situation.

Then, to avoid having such big files in the repository, build the
splashes from SVGs using imagemagick. This makes it easier for people to
modify the splashes.

And finally, add a file to configure the progress bar for updates, one
for each device. This lets us have progress bars in appropriate places
with appropriate sizes.